requirements.txt
aliyun-python-sdk-alidns==2.0.10
aliyun-python-sdk-core==2.13.5
aliyun-python-sdk-core-v3==2.13.3
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.1
django-crontab==0.7.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
dnspython==1.16.0
idna==2.8
jmespath==0.9.4
mysqlclient==1.4.4
Pillow==6.1.0
pycparser==2.19
pycryptodome==3.8.2
pyOpenSSL==19.0.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
